requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.5
openai
faiss-cpu
beautifulsoup4
requests
python-dotenv
sentence-transformers
fastapi
uvicorn